refactor(employee): Employee 관련 클래스 user로 이동
회원 관련 클래스인 Employee 관련 클래스들을 user 디렉토리로 옮겼습니다.
이제 회원 관련 기능은 user 디렉토리 안에서 확인할 수 있습니다.
또한 불필요하게 같은 테이블에 대한 Entity가 여러 개 존재하던 것을 하나의 Entity로 관리하게 되었습니다.
