--- src/main/java/org/triumers/kmsback/employee/command/Application/controller/CmdEmployeeController.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.command.Application.controller;
-
-public class CmdEmployeeController {
-}
