--- src/main/java/org/triumers/kmsback/employee/query/aggregate/entity/QryEmployee.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.query.aggregate.entity;
-
-public class QryEmployee {
-}
