--- src/main/java/org/triumers/kmsback/employee/query/aggregate/vo/QryResponseMessageVO.java
@@ -1,5 +1,0 @@
-package org.triumers.kmsback.employee.query.aggregate.vo;
-
-public class QryResponseMessageVO {
-    private String message;
-}
