--- src/main/java/org/triumers/kmsback/employee/query/controller/QryEmployeeController.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.query.controller;
-
-public class QryEmployeeController {
-}
