--- src/main/java/org/triumers/kmsback/employee/query/dto/QryEmployeeDTO.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.query.dto;
-
-public class QryEmployeeDTO {
-}
