--- src/main/java/org/triumers/kmsback/employee/query/repository/QryEmployeeRepository.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.query.repository;
-
-public class QryEmployeeRepository {
-}
