--- src/main/java/org/triumers/kmsback/employee/query/service/QryEmployeeService.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.query.service;
-
-public interface QryEmployeeService {
-}
