--- src/main/java/org/triumers/kmsback/employee/query/service/QryEmployeeServiceImpl.java
@@ -1,4 +1,0 @@
-package org.triumers.kmsback.employee.query.service;
-
-public class QryEmployeeServiceImpl implements QryEmployeeService {
-}
