+++ src/main/java/org/triumers/kmsback/post/command/Application/service/CmdPostServiceImpl.java
@@ -2,11 +2,11 @@ package org.triumers.kmsback.post.command.Application.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.triumers.kmsback.auth.command.Application.service.AuthService;
-import org.triumers.kmsback.auth.command.domain.aggregate.entity.Auth;
+import org.triumers.kmsback.user.command.Application.service.AuthService;
+import org.triumers.kmsback.user.command.domain.aggregate.entity.Employee;
 import org.triumers.kmsback.common.exception.NotAuthorizedException;
 import org.triumers.kmsback.common.exception.NotLoginException;
-import org.triumers.kmsback.employee.command.Application.service.CmdEmployeeService;
+import org.triumers.kmsback.user.command.Application.service.CmdEmployeeService;
 import org.triumers.kmsback.post.command.Application.dto.*;
 import org.triumers.kmsback.post.command.domain.aggregate.entity.*;
 import org.triumers.kmsback.post.command.domain.repository.*;
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.triumers.kmsback.auth.command.domain.aggregate.enums.UserRole.ROLE_ADMIN;
+import static org.triumers.kmsback.user.command.domain.aggregate.enums.UserRole.ROLE_ADMIN;
 
 @Service
 public class CmdPostServiceImpl implements CmdPostService {
@@ -46,7 +46,7 @@ public class CmdPostServiceImpl implements CmdPostService {
     @Override
     public CmdPostAndTagsDTO registPost(CmdPostAndTagsDTO post) throws NotLoginException {
 
-        Auth employee = authService.whoAmI();
+        Employee employee = authService.whoAmI();
         CmdPost registPost = new CmdPost(post.getId(), post.getTitle(), post.getContent(), post.getCreatedAt(),
                 employee.getId(), post.getOriginId(), post.getTabRelationId(), post.getCategoryId());
 
@@ -125,7 +125,7 @@ public class CmdPostServiceImpl implements CmdPostService {
     @Override
     public Boolean isAuthorizedToPost(int originId) throws NotLoginException {
 
-        Auth employee = authService.whoAmI();
+        Employee employee = authService.whoAmI();
         int authorId = cmdPostRepository.findAuthorIdById(originId);
 
         if (employee.getId() == authorId || employee.getUserRole() == ROLE_ADMIN) {
@@ -137,7 +137,7 @@ public class CmdPostServiceImpl implements CmdPostService {
     @Override
     public CmdLikeDTO likePost(CmdLikeDTO like) throws NotLoginException {
 
-        Auth employee = authService.whoAmI();
+        Employee employee = authService.whoAmI();
         try {
             CmdLike likePost = cmdLikeRepository.findByEmployeeIdAndPostId(employee.getId(), like.getPostId());
 
@@ -162,7 +162,7 @@ public class CmdPostServiceImpl implements CmdPostService {
     @Override
     public CmdFavoritesDTO favoritePost(CmdFavoritesDTO favorite) throws NotLoginException {
 
-        Auth employee = authService.whoAmI();
+        Employee employee = authService.whoAmI();
         try {
             CmdFavorites favoritePost = cmdFavoritesRepository.findByEmployeeIdAndPostId(employee.getId(), favorite.getPostId());
             if (favoritePost != null) {  // unfavorite
