+++ src/main/java/org/triumers/kmsback/post/query/controller/QryPostController.java
@@ -6,8 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.triumers.kmsback.employee.command.Application.dto.CmdEmployeeDTO;
-import org.triumers.kmsback.employee.query.dto.QryEmployeeDTO;
+import org.triumers.kmsback.user.command.Application.dto.CmdEmployeeDTO;
 import org.triumers.kmsback.post.query.aggregate.vo.QryRequestPost;
 import org.triumers.kmsback.post.query.dto.QryPostAndTagsDTO;
 import org.triumers.kmsback.post.query.service.QryPostService;
