+++ src/main/java/org/triumers/kmsback/post/query/service/QryPostService.java
@@ -2,11 +2,8 @@ package org.triumers.kmsback.post.query.service;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.triumers.kmsback.employee.command.Application.dto.CmdEmployeeDTO;
-import org.triumers.kmsback.employee.query.dto.QryEmployeeDTO;
-import org.triumers.kmsback.post.query.aggregate.entity.QryPostAndTag;
+import org.triumers.kmsback.user.command.Application.dto.CmdEmployeeDTO;
 import org.triumers.kmsback.post.query.aggregate.vo.QryRequestPost;
-import org.triumers.kmsback.post.query.dto.QryLikeDTO;
 import org.triumers.kmsback.post.query.dto.QryPostAndTagsDTO;
 
 import java.util.List;
