+++ src/main/java/org/triumers/kmsback/post/query/service/QryPostServiceImpl.java
@@ -4,20 +4,13 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.triumers.kmsback.employee.command.Application.dto.CmdEmployeeDTO;
-import org.triumers.kmsback.employee.command.Application.service.CmdEmployeeService;
-import org.triumers.kmsback.employee.query.dto.QryEmployeeDTO;
-import org.triumers.kmsback.employee.query.service.QryEmployeeService;
-import org.triumers.kmsback.post.command.Application.dto.CmdTagDTO;
-import org.triumers.kmsback.post.command.domain.aggregate.entity.CmdTag;
+import org.triumers.kmsback.user.command.Application.dto.CmdEmployeeDTO;
+import org.triumers.kmsback.user.command.Application.service.CmdEmployeeService;
 import org.triumers.kmsback.post.query.aggregate.entity.QryLike;
 import org.triumers.kmsback.post.query.aggregate.entity.QryPostAndTag;
 import org.triumers.kmsback.post.query.aggregate.entity.QryTag;
 import org.triumers.kmsback.post.query.aggregate.vo.QryRequestPost;
-import org.triumers.kmsback.post.query.dto.QryLikeDTO;
 import org.triumers.kmsback.post.query.dto.QryPostAndTagsDTO;
-import org.triumers.kmsback.post.query.dto.QryPostDTO;
-import org.triumers.kmsback.post.query.dto.QryTagDTO;
 import org.triumers.kmsback.post.query.repository.QryPostMapper;
 
 import java.util.ArrayList;
