+++ src/main/java/org/triumers/kmsback/user/command/Application/service/CmdEmployeeServiceImpl.java
@@ -1,32 +1,32 @@
-package org.triumers.kmsback.employee.command.Application.service;
+package org.triumers.kmsback.user.command.Application.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.triumers.kmsback.employee.command.Application.dto.CmdEmployeeDTO;
-import org.triumers.kmsback.employee.command.domain.aggregate.entity.CmdEmployee;
-import org.triumers.kmsback.employee.command.domain.repository.CmdEmployeeRepository;
+import org.triumers.kmsback.user.command.Application.dto.CmdEmployeeDTO;
+import org.triumers.kmsback.user.command.domain.aggregate.entity.Employee;
+import org.triumers.kmsback.user.command.domain.repository.EmployeeRepository;
 
 import java.util.List;
 
 @Service
 public class CmdEmployeeServiceImpl implements CmdEmployeeService {
 
-    private final CmdEmployeeRepository cmdEmployeeRepository;
+    private final EmployeeRepository employeeRepository;
 
     @Autowired
-    public CmdEmployeeServiceImpl(CmdEmployeeRepository cmdEmployeeRepository) {
-        this.cmdEmployeeRepository = cmdEmployeeRepository;
+    public CmdEmployeeServiceImpl(EmployeeRepository employeeRepository) {
+        this.employeeRepository = employeeRepository;
     }
 
     @Override
     public CmdEmployeeDTO findEmployeeById(int id) {
 
-        return employeeToEmployeeDTO(cmdEmployeeRepository.findById(id));
+        return employeeToEmployeeDTO(employeeRepository.findById(id));
     }
 
     @Override
     public CmdEmployeeDTO findEmployeeByEmail(String email) {
-        return employeeToEmployeeDTO(cmdEmployeeRepository.findByEmail(email));
+        return employeeToEmployeeDTO(employeeRepository.findByEmail(email));
     }
 
     @Override
@@ -35,7 +35,7 @@ public class CmdEmployeeServiceImpl implements CmdEmployeeService {
         return null;
     }
 
-    private CmdEmployeeDTO employeeToEmployeeDTO(CmdEmployee employee) {
+    private CmdEmployeeDTO employeeToEmployeeDTO(Employee employee) {
         return new CmdEmployeeDTO(employee.getId(), employee.getEmail(), employee.getName(),
                 employee.getProfileImg(), employee.getUserRole().name(), employee.getStartDate(), employee.getEndDate(),
                 employee.getPhoneNumber(), employee.getTeamId(), employee.getPositionId(), employee.getRankId());
