+++ src/main/java/org/triumers/kmsback/user/command/domain/repository/EmployeeRepository.java
@@ -6,5 +6,8 @@ import org.triumers.kmsback.user.command.domain.aggregate.entity.Employee;
 
 @Repository
 public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+
+    Employee findById(int id);
+
     Employee findByEmail(String email);
 }
